requirements.txt
python-telegram-bot==13.15
requests==2.31.0
Flask==2.3.3
gunicorn==21.2.0